requirements.txt
discord.py
python-dotenv
supabase
supabase_auth
supabase_functions
asyncio
aiohttp
psycopg2-binary